I've made progress on implementing internationalization for the App Router:
- I removed the unsupported i18n object from `next.config.mjs`.
- I added middleware for locale detection (en, fr), defaulting to en.
- I restructured the app directory to `src/app/[lang]`.
- I created `en.json` and `fr.json` dictionaries.
- I implemented a `getDictionary` function to load translations.
- I updated the root layout and main page to use i18n.
- I added `generateStaticParams` for en and fr locales.
