+++ package.json
@@ -16,6 +16,7 @@
 		"test:watch": "jest --watch"
 	},
 	"dependencies": {
+		"@formatjs/intl-localematcher": "^0.6.1",
 		"@radix-ui/react-slot": "1.2.3",
 		"@react-three/drei": "10.1.2",
 		"@react-three/fiber": "9.1.2",
@@ -27,6 +28,7 @@
 		"github-slugger": "2.0.0",
 		"highlight.js": "11.11.1",
 		"lucide-react": "0.511.0",
+		"negotiator": "^1.0.0",
 		"next": "15.3.3",
 		"next-sitemap": "4.2.3",
 		"next-themes": "0.4.6",
