+++ src/app/[lang]/artist/layout.jsx
