+++ src/app/[lang]/artist/page.jsx
