+++ src/app/[lang]/blog/BlogClient.jsx
