+++ src/app/[lang]/blog/articles/[slug]/BlogPageClient.jsx
