+++ src/app/[lang]/blog/articles/[slug]/layout.jsx
