+++ src/app/[lang]/blog/articles/[slug]/page.jsx
