+++ src/app/[lang]/blog/categories/[slug]/page.jsx
