+++ src/app/[lang]/blog/layout.jsx
@@ -1,6 +1,6 @@
-import siteMetadata from '../../../utils/siteMetaData'
+import siteMetadata from '@/utils/siteMetaData'
 
-import { cx } from '/src/utils'
+import { cx } from '@/utils'
 
 export const revalidate = 60
