+++ src/app/[lang]/experiment/slime/layout.jsx
