+++ src/app/[lang]/experiment/slime/page.jsx
