+++ src/app/[lang]/favicon.ico
