+++ src/app/[lang]/favicon.png
