+++ src/app/[lang]/opengraph-image.jsx
