+++ src/app/[lang]/projects/ProjectClient.jsx
