+++ src/app/[lang]/projects/[id]/page.jsx
