+++ src/dictionaries/en.json
@@ -1,0 +1,3 @@
+{
+  "greeting": "Hello!"
+}
