+++ src/dictionaries/fr.json
@@ -1,0 +1,3 @@
+{
+  "greeting": "Bonjour!"
+}
