+++ src/middleware.js
@@ -1,0 +1,40 @@
+import { NextResponse } from 'next/server';
+import { match } from '@formatjs/intl-localematcher';
+import Negotiator from 'negotiator';
+
+const locales = ['fr', 'en'];
+const defaultLocale = 'en';
+
+function getLocale(request) {
+  const headers = {};
+  request.headers.forEach((value, key) => (headers[key] = value));
+
+  const languages = new Negotiator({ headers }).languages();
+  
+  try {
+    return match(languages, locales, defaultLocale);
+  } catch (error) {
+    // Fallback to defaultLocale if match throws an error (e.g., no suitable match)
+    return defaultLocale;
+  }
+}
+
+export function middleware(request) {
+  const { pathname } = request.nextUrl;
+
+  const pathnameHasLocale = locales.some(
+    (locale) => pathname.startsWith(`/${locale}/`) || pathname === `/${locale}`
+  );
+
+  if (pathnameHasLocale) return;
+
+  const locale = getLocale(request);
+  request.nextUrl.pathname = `/${locale}${pathname}`;
+  return NextResponse.redirect(request.nextUrl);
+}
+
+export const config = {
+  matcher: [
+    '/((?!api|_next/static|_next/image|favicon.ico|.*\\.(?:svg|png|jpg|jpeg|gif|webp)$).*)',
+  ],
+};
